Imprime los primeros digitos que se aparescan con su signo si es negativo

+++ pointers_arrays_strings/100-atoi.c
@@ -7,7 +7,7 @@
 int _atoi(char *s)
 {
 	int i = 1;
-	long int a = 0;
+	int a = 0;
 
 	do
 	{
